[td-225]fix compiler error.

+++ src/query/src/qExecutor.c
@@ -4607,7 +4607,7 @@ SOperatorInfo *createMultiwaySortOperatorInfo(SQueryRuntimeEnv *pRuntimeEnv, SEx
       len += pExpr[i].base.colBytes;
     }
 
-    int32_t numOfCols = (pInfo->orderColumnList != NULL)? taosArrayGetSize(pInfo->orderColumnList):0;
+    int32_t numOfCols = (pInfo->orderColumnList != NULL)? (int32_t) taosArrayGetSize(pInfo->orderColumnList):0;
     pInfo->prevRow = calloc(1, (POINTER_BYTES * numOfCols + len));
     int32_t offset = POINTER_BYTES * numOfCols;
